Reorganized recycle bin and user tests
