+++ 3.Tests/UI/Features/RecycleBin/RecycleBin.feature
@@ -1,4 +1,4 @@
-Feature: Empty recycle bin
+Feature: Recycle Bin
     As a logged in user, the user should be able to empty the recible bin
     Background:
         Given the user is logged in
